+++ packages/shared/src/validators/holdout.ts
@@ -92,14 +92,3 @@ export type HoldoutInterface = z.infer<typeof holdoutValidator>;
 export type HoldoutInterfaceStringDates = z.infer<
   typeof _holdoutStringDatesValidator
 >;
-
-// Capped at 0.5 because an equal control group is reserved alongside the holdout.
-export const MAX_HOLDOUT_SIZE = 0.5;
-
-export function holdoutSizeToCoverage(holdoutSize: number): number {
-  return holdoutSize * 2;
-}
-
-export function coverageToHoldoutSize(coverage: number): number {
-  return coverage / 2;
-}
